change argo cd instance label to prevent conflict with helm
argocd uses app.kubernetes.io/instance to keep track of each app's resources.
But this can conflict with the values that helm uses. In this case, argo set the instance to "nginx-ingress" which caused the ServiceMonitor labels not to match, leading to the issues with scraping metrics. See https://github.com/argoproj/argo-cd/issues/1482